feat: Add BattleCharacterSequences property to BattleCharacter constructor
This commit adds the `BattleCharacterSequences` property to the `BattleCharacter` constructor in the `BattleCharacter.cs` file. The property allows passing a `BattleCharacterSequences` object to the constructor, enabling the player character to play sequences when taking damage. This change enhances the gameplay experience by adding visual effects to indicate damage taken by the player.