feat: Remove unused AilmentSequence asset and add new OnTurnStart asset
This commit removes the unused AilmentSequence asset "101016.OnCalculateRecoveryBehaviourPoint" and adds a new asset "101016.OnTurnStart" in the SoulRPG project. The removed asset was no longer needed and was deleted. The new asset represents a sequence of actions to be executed when a battle character's turn starts. This change improves the organization of the AilmentSequences and enhances the functionality of the game.

Note: The commit message has been generated based on the provided code changes and recent commits.