feat: Add new asset for skill action 101041
This commit adds a new asset file, 101041.asset, to the SoulRPG project. The code change includes the creation of the asset file, which defines the skill action with the ID 101041. This addition enhances the functionality of the game by introducing a new skill action for characters to perform.

Note: The commit message has been generated based on the provided code changes and recent commits.