feat: Add new Ailment prefab
This commit adds a new Ailment prefab file to the SoulRPG project. The code change includes the creation of the prefab file, which defines the UI for displaying information about ailments in the game menu. This addition enhances the functionality of the game by providing a visual representation of ailments and their effects.

Note: The commit message has been generated based on the provided code changes and recent commits.