feat: Add new document to UIBundle.Game.Menu.prefab
This commit adds a new document to the UIBundle.Game.Menu.prefab file in the SoulRPG project. The code change includes the addition of a new document entry in the prefab, which enhances the functionality of the game menu. This modification improves the organization and structure of the codebase.

Note: The commit message has been generated based on the provided code changes and recent commits.