feat: Add RecoveryHitPointAsync method to BattleCharacter.cs
This commit adds a new method, RecoveryHitPointAsync, to the BattleCharacter.cs file in the SoulRPG project. The RecoveryHitPointAsync method is responsible for recovering hit points for a character. It takes an integer parameter, recovery, which represents the amount of hit points to be recovered. The method calls the RecoveryHitPoint method of the BattleStatus class to perform the recovery. This addition enhances the battle system by providing a convenient way to recover hit points.

Note: The commit message has been generated based on the provided code changes and recent commits.