feat: Update thumbnail images for weapons 101010, 101020, and 101025
This commit updates the thumbnail images for weapons 101010, 101020, and 101025 in the MasterData.asset file. The previous thumbnails were empty, but now they are set to the appropriate images. This change improves the visual representation of the weapons in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.