feat: Update GameItemInformationView to display weapon attack attribute
This commit updates the GameItemInformationView in the SoulRPG project to display the attack attribute of a weapon. The code change adds a line that sets the text of the "Value" TMP_Text component in the "Weapon.Attribute" HKUIDocument to the UI string representation of the weapon's attack attribute. This enhancement improves the user interface by providing more detailed information about the weapon's attributes.

Note: The commit message has been generated based on the provided code changes and recent commits.