feat: Add OnDeselectSequences script for button systems
This commit adds a new script, OnDeselectSequences.cs, to the ButtonSystems folder in the SoulRPG project. The script handles the deselection event of a selectable UI element and plays a sequence asynchronously using the SequenceMonobehaviour component. This addition enhances the functionality of the button systems in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.