feat: Add Weapon.29.png and Weapon.30.png textures
This commit adds the Weapon.29.png and Weapon.30.png textures to the SoulRPG project. These textures are used for rendering weapons in the game. The addition of these textures enhances the visual quality and realism of the weapons.

Note: The commit message has been generated based on the provided code changes and recent commits.