feat: Update OnTakeDamageAsync method parameters in AilmentController and AilmentElement
The OnTakeDamageAsync method in the AilmentController and AilmentElement classes now accept an additional parameter, DamageData, to provide more information about the damage taken. This change ensures that the damage data is properly passed to the elements when handling damage events.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.