feat: Add DamageData struct to BattleSystems
This commit adds the `DamageData` struct to the `BattleSystems` namespace. The `DamageData` struct contains two properties: `Damage` of type `int` and `AttackAttribute` of type `Define.AttackAttribute`. This struct is used to store information about damage and attack attributes in the battle systems.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.