feat: Add None attack attribute to Define.cs
This commit adds the None attack attribute to the Define.cs file in the SoulRPG/Scripts directory. The None attack attribute is used to represent a lack of attack attribute. This addition allows for more flexibility in defining attack attributes in the game.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.