feat: Update AilmentSequences/101018.OnTakeDamage.Success.asset
This commit adds a new asset file, 101018.OnTakeDamage.Success.asset, to the AilmentSequences directory. The asset contains a MonoBehaviour with sequences and references related to taking damage. This addition enhances the functionality of the AilmentSequences system.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.