refactor: Update GiveDamage.cs to use attackAttribute from weapon if available
This commit refactors the GiveDamage.cs file to use the attackAttribute property from the weapon if it is available. This ensures that the correct attack attribute is used when creating the damageData struct. By checking if the weapon is null and using the null coalescing operator, we can safely assign the attackAttribute value to the damageData struct. This change improves the accuracy of damage calculations in the game.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.