feat: Remove "Damage" from container in GiveDamage.cs
This commit removes the "Damage" key from the container in the GiveDamage.cs file. This change is necessary to update the logic related to damage calculation. By removing the key, the code ensures that the correct value is registered for the "Damage" key in the container. This modification improves the accuracy of damage calculations in the SoulRPG game.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.