feat: Update ScrollScanShader.Canvas1 material
This commit adds a new material file, ScrollScanShader.Canvas1.mat, to the project. The material is used for the ScrollScanShader effect in the EasyCustomVFXUI package. This addition enhances the visual effects of the UI elements in the game.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.