feat: Add new prefab for KnifeCount UI element
This commit adds a new prefab file, UIElement.KnifeCount.prefab, to the project. The prefab is used for displaying the knife count UI element in the game. This addition enhances the user interface by providing a visual representation of the number of knives available to the player.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.